+++ src/game/blocks/normalBlock.ts
@@ -4,5 +4,6 @@ export class NormalBlock extends Block {
     constructor(x: number, y: number, w: number, h: number) {
         super(x, y, w, h);
         this.type = 'block';
+        this.texture = 'block';
     }
 }